+++ frontend/src/App.jsx
@@ -11,6 +11,7 @@ import OrderPage from './pages/OrderPage'
 import MarketBoardPage from './pages/MarketBoardPage'
 import MacroPage from './pages/MacroPage'
 import AdvisorPage from './pages/AdvisorPage'
+import PortfolioPage from './pages/PortfolioPage'
 import ToastNotification from './components/common/ToastNotification'
 import { useNotification } from './hooks/useNotification'
 
@@ -48,6 +49,7 @@ export default function App() {
           <Route path="/market-board" element={<MarketBoardPage />} />
           <Route path="/macro" element={<MacroPage />} />
           <Route path="/advisor" element={<AdvisorPage notify={notify} />} />
+          <Route path="/portfolio" element={<PortfolioPage />} />
         </Routes>
       </main>
       <ToastNotification toasts={toasts} removeToast={removeToast} />